Requested changes - change package from resources to resource - renamed files accordingly - changed from if statement to switch statement. - deleted empty README.md
