+++ internal/scout/resource/resource_test.go
@@ -1,4 +1,4 @@
-package resources
+package resource
 
 import (
 	"context"
